refactor: main 컴포넌트를 barrel 방식으로 변경하기
