+++ app/my-profile/_layout.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Stack } from "expo-router";
 import { StatusBar } from "expo-status-bar";
 import CreateHeader from "@/components/my-profile/create/header";
-import MainHeader from "@/components/my-profile/main/header";
+import { Header as MainHeader } from "@/components/my-profile/main";
 import { Header as ActivityHeader } from "@/components/my-profile/activity";
 
 const ProfileLayout = () => {
