+++ components/my-profile/main/index.ts
@@ -1,0 +1,7 @@
+import ControlButtons from "./control-buttons";
+import Header from "./header";
+
+export {
+  ControlButtons,
+  Header,
+}
